+++ src/lib.rs
@@ -88,8 +88,9 @@ pub use friday::{
     FridayGenerationSummary, FridayLocalCheckKind, FridayLocalCheckStatus,
     FridayBrowserVerificationReport, FridayBrowserVerificationTarget, FridayDeployGate,
     FridayLocalExecutionCheck, FridayLocalExecutionReport, FridayMemoryRecord, FridayMemoryState,
-    FridayMultimodalRuntimeRecord, FridayMultimodalSurface, FridayPermissionFinding,
-    FridayOcrSmokeReport, FridayOcrSmokeStatus, FridayPermissionScope, FridayPreviewRunner,
+    FridayMultimodalModelFile, FridayMultimodalRuntimeRecord, FridayMultimodalSurface,
+    FridayOcrSmokeReport, FridayOcrSmokeStatus, FridayPermissionFinding, FridayPermissionScope,
+    FridayPreviewRunner,
     FridayProductPlan, FridayProjectFile, FridayProjectRecord, FridayResearchEventKind,
     FridayResearchExportManifest,
     FridayResearchReport, FridayResearchRunEvent, FridayResearchStage, FridayResearchStageKind,
@@ -101,10 +102,11 @@ pub use friday::{
     FridayUiStateTone,
     FridayVoiceRuntimeRecord, FridayWorkspaceArea, FridayWorkspaceManifest,
     FridayWorkspaceSnapshot, FridayWorkspaceStore, FridayWorkspaceView, MetasearchExecutionMode,
-    FridayVerificationStatus, MetasearchExecutionTarget,
+    FridayVerificationStatus, FridayVlmContractReport, FridayVlmContractStatus,
+    MetasearchExecutionTarget,
     default_friday_browser_verification_report, default_friday_local_execution_checks,
     default_friday_product_plan, default_friday_ui_integration_plan, friday_answer_search_plan,
-    friday_research_search_plan, run_friday_ocr_smoke,
+    friday_research_search_plan, run_friday_ocr_smoke, run_friday_vlm_contract,
 };
 pub use long_context::{LongContextExecutionPlan, LongContextTask, RlmBridge};
 pub use models::{KokoroTTS, LocalLlm, LocalSttEngine, MoonshineSTT};